requirements.txt
streamlit
ffmpeg-python
moviepy
pydub
openai-whisper
numpy
torch